multiple code reorganize and fixes
- migrated all graph-specific code to a new module `diagram_gen`
- Set display name etc. as the class attribute of metrics
- auto calculate layout based on number of datasets passed in
- separated summarizing function and plotting function
- minor fixes
