+++ evaluation/evaluate_metrics/builtin_metrics.py
@@ -18,6 +18,10 @@ class BuiltinRatingMetric(BaseMetric):
 
 class BuiltinRelevanceMetric(BuiltinRatingMetric):
     METRIC_NAME = "gpt_relevance"
+    
+    DISPLAY_TITLE = "GPT Relevance Rating"
+    SHORT_NAME = "Relevance"
+    Y_AXIS_LABEL = "Rating Score (1-5)"
 
     @classmethod
     def evaluator_fn(cls, openai_config, **kwargs):
@@ -27,6 +31,10 @@ class BuiltinRelevanceMetric(BuiltinRatingMetric):
 class BuiltinCoherenceMetric(BuiltinRatingMetric):
     METRIC_NAME = "gpt_coherence"
 
+    DISPLAY_TITLE = "GPT Coherence Rating"
+    SHORT_NAME = "Coherence"
+    Y_AXIS_LABEL = "Rating Score (1-5)"
+
     @classmethod
     def evaluator_fn(cls, openai_config, **kwargs):
         return CoherenceEvaluator(openai_config)
@@ -35,6 +43,10 @@ class BuiltinCoherenceMetric(BuiltinRatingMetric):
 class BuiltinGroundednessMetric(BuiltinRatingMetric):
     METRIC_NAME = "gpt_groundedness"
 
+    DISPLAY_TITLE = "GPT Groundedness Rating"
+    SHORT_NAME = "Groundedness"
+    Y_AXIS_LABEL = "Rating Score (1-5)"
+
     @classmethod
     def evaluator_fn(cls, openai_config, **kwargs):
         return GroundednessEvaluator(openai_config)
@@ -42,6 +54,10 @@ class BuiltinGroundednessMetric(BuiltinRatingMetric):
 
 class BuiltinSimilarityMetric(BuiltinRatingMetric):
     METRIC_NAME = "gpt_similarity"
+    
+    DISPLAY_TITLE = "GPT Similarity Rating"
+    SHORT_NAME = "Similarity"
+    Y_AXIS_LABEL = "Rating Score (1-5)"
 
     @classmethod
     def evaluator_fn(cls, openai_config, **kwargs):
@@ -50,6 +66,10 @@ class BuiltinSimilarityMetric(BuiltinRatingMetric):
 
 class BuiltinFluencyMetric(BuiltinRatingMetric):
     METRIC_NAME = "gpt_fluency"
+    
+    DISPLAY_TITLE = "GPT Fluency Rating"
+    SHORT_NAME = "Fluency"
+    Y_AXIS_LABEL = "Rating Score (1-5)"
 
     @classmethod
     def evaluator_fn(cls, openai_config, **kwargs):
@@ -58,6 +78,10 @@ class BuiltinFluencyMetric(BuiltinRatingMetric):
 
 class BuiltinF1ScoreMetric(BaseMetric):
     METRIC_NAME = "f1_score"
+    
+    DISPLAY_TITLE = "F1 Score"
+    SHORT_NAME = "F1 Score"
+    Y_AXIS_LABEL = "F1 Score"
 
     @classmethod
     def evaluator_fn(cls, **kwargs):
