+++ evaluation/evaluate_metrics/code_metrics.py
@@ -9,6 +9,10 @@ logger = logging.getLogger("evaluation")
 class AnswerLengthMetric(BaseMetric):
     METRIC_NAME = "answer_length"
 
+    DISPLAY_TITLE = "Answer Length"
+    SHORT_NAME = "Length"
+    Y_AXIS_LABEL = "Length by Word Count"
+
     @classmethod
     def evaluator_fn(cls, **kwargs):
         def answer_length(*, answer, **kwargs):
@@ -32,6 +36,9 @@ class AnswerLengthMetric(BaseMetric):
 
 class HasCitationMetric(BaseMetric):
     METRIC_NAME = "has_citation"
+    
+    DISPLAY_TITLE = "Has Citation"
+    SHORT_NAME = "Has Citation"
 
     @classmethod
     def evaluator_fn(cls, **kwargs):
@@ -54,6 +61,9 @@ class HasCitationMetric(BaseMetric):
 
 class CitationMatchMetric(BaseMetric):
     METRIC_NAME = "citation_match"
+    
+    DISPLAY_TITLE = "Citation Match"
+    SHORT_NAME = "Citation Match"
 
     @classmethod
     def evaluator_fn(cls, **kwargs):
@@ -80,6 +90,10 @@ class CitationMatchMetric(BaseMetric):
 
 class LatencyMetric(BaseMetric):
     METRIC_NAME = "latency"
+    
+    DISPLAY_TITLE = "Response Latency (seconds)"
+    SHORT_NAME = "Latency"
+    Y_AXIS_LABEL = "Latency in Seconds"
 
     @classmethod
     def evaluator_fn(cls, **kwargs):
