+++ tests/test_evaluate.py
@@ -125,49 +125,50 @@ def test_run_evaluation():
         results_dir = Path(tempdir) / "results"
 
         with mock.patch("evaluation.evaluate.load_jsonl", return_value=[{"question": "What is 2 + 2?", "truth": "4"}]):
-            with mock.patch("evaluation.evaluate.summarize_results_and_plot"):
-                with mock.patch("evaluation.evaluate.service_setup.get_openai_config", return_value={}):
-                    with mock.patch(
-                        "evaluation.evaluate.send_question_to_target",
-                        return_value={"answer": "4", "context": "2 + 2 = 4", "latency": 1.0},
-                    ):
+            with mock.patch("evaluation.evaluate.dump_summary", return_value=None):
+                with mock.patch("evaluation.evaluate.plot_diagrams", return_value=None):
+                    with mock.patch("evaluation.evaluate.service_setup.get_openai_config", return_value={}):
+                        with mock.patch(
+                            "evaluation.evaluate.send_question_to_target",
+                            return_value={"answer": "4", "context": "2 + 2 = 4", "latency": 1.0},
+                        ):
 
-                        metrics_by_name["mock_metric"] = type(
-                            "MockMetric",
-                            (),
-                            {
-                                "METRIC_NAME": "mock_metric",
-                                "evaluator_fn": staticmethod(
-                                    lambda openai_config: lambda question, answer, context, ground_truth: {
-                                        "mock_metric_score": 3.0
-                                    }
-                                ),
-                                "get_aggregate_stats": staticmethod(
-                                    lambda df, passing_rate: {"pass_rate": 0.67, "mean_rating": 3.0}
-                                ),
-                            },
-                        )
+                            metrics_by_name["mock_metric"] = type(
+                                "MockMetric",
+                                (),
+                                {
+                                    "METRIC_NAME": "mock_metric",
+                                    "evaluator_fn": staticmethod(
+                                        lambda openai_config: lambda question, answer, context, ground_truth: {
+                                            "mock_metric_score": 3.0
+                                        }
+                                    ),
+                                    "get_aggregate_stats": staticmethod(
+                                        lambda df, passing_rate: {"pass_rate": 0.67, "mean_rating": 3.0}
+                                    ),
+                                },
+                            )
 
-                        openai_config = AzureOpenAIModelConfiguration("azure")
-                        openai_config.model = "mock_model"
-                        target_url = "http://mock-target-url.com"
-                        passing_rate = 3
-                        max_workers = 2
-                        target_parameters = {}
-                        requested_metrics = ["mock_metric"]
+                            openai_config = AzureOpenAIModelConfiguration("azure")
+                            openai_config.model = "mock_model"
+                            target_url = "http://mock-target-url.com"
+                            passing_rate = 3
+                            max_workers = 2
+                            target_parameters = {}
+                            requested_metrics = ["mock_metric"]
 
-                        success = run_evaluation(
-                            openai_config=openai_config,
-                            testdata_path=testdata_path,
-                            results_dir=results_dir,
-                            target_url=target_url,
-                            passing_rate=passing_rate,
-                            max_workers=max_workers,
-                            target_parameters=target_parameters,
-                            requested_metrics=requested_metrics,
-                        )
+                            success = run_evaluation(
+                                openai_config=openai_config,
+                                testdata_path=testdata_path,
+                                results_dir=results_dir,
+                                target_url=target_url,
+                                passing_rate=passing_rate,
+                                max_workers=max_workers,
+                                target_parameters=target_parameters,
+                                requested_metrics=requested_metrics,
+                            )
 
-                        assert success
+                            assert success
 
 
 class MockResponse:
